+++ nequip/train/_loss.py
@@ -4,7 +4,7 @@ import torch.nn
 from torch_scatter import scatter
 
 from nequip.data import AtomicDataDict
-from nequip.utils.auto_init import instantiate_from_cls_name
+from nequip.utils import instantiate_from_cls_name
 
 
 class SimpleLoss:
